Package/Scripts - Add ability to calculate shade for a topo.
Add the ability to calculate shading for a topo based on sun angle. This
copies the `sunang` functions from SMRF and a subset of the `load_topo`
class to ensure compatibility with topocalc.

Copied the classes instead of adding SMRF as a package dependency as
SMRF has its own set of required packages which are not needed here.